ppc: Rework CPU compatibility testing across migration
Migrating between different CPU versions is quite complicated for ppc.
A long time ago, we ensure identical CPU versions at either end by checking
the PVR had the same value.  However, this breaks under KVM HV, because
we always have to use the host's PVR - it's not virtualized.  That would
mean we couldn't migrate between hosts with different PVRs, even if the
CPUs are close enough to compatible in practice (sometimes identical cores
with different surrounding logic have different PVRs, so this happens in
practice quite often).

So, we removed the PVR check, but instead checked that several flags
indicating supported instructions matched.  This turns out to be a bad
idea, because those instruction masks are not architected information, but
essentially a TCG implementation detail.  So changes to qemu internal CPU
modelling can break migration - this happened between qemu-2.6 and
qemu-2.7.

Modern server-class CPUs can be placed into compatibility modes.  Now that
we're handling those properly, we finally have the information to sanely
deal with CPU compatibility across migration.

This patch bumps the migration version number for the ppc CPU removing the
instruction mask field (and some other unwise VMSTATE_EQUAL checks), and
adding the compatibility PVR to the migration stream.

We consider the CPUs compatible for migration if:
    * The source was running in a compatibility mode which the destination
      supports
OR  * The source has a PVR matching the same qemu CPU class as the
      destination, either an exact match or an approximate match determined
      by the cpu class's pvr_match hook.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>