spapr_pci: Allow PCI-Express devices
Currently the PCI host bridge (PHB) for the pseries machine type always
creates a plain PCI bus (TYPE_PCI_BUS).  This means that qemu won't allow
PCIe devices to be attached, and consequently, PCIe devices can't be used
on pseries.

This limitation isn't inherent in the PAPR specification.  In fact, because
access to the PCI bus is paravirtualized via hypercalls, from the guest
point of view there is very little difference between a plain PCI and PCIe
bus under PAPR.

So, to allow PCIe devices, add a "pcie" option to the PHB device, which
when enabled creates a PCIe bus.  In addition, on PCIe devices under such
a PHB, we publish the device tree flag which indicates that the guest may
access PCIe extended config space (via RTAS calls which we've already
implemented).

Note that this doesn't change the default PHB to PCIe.  We probably want to
do that in future, but we need to sort out some possible breakages with
libvirt and existing users first.  This at least lets knowledgeable users
experiment with PCIe devices in the meantime.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>