spapr: Change PCI bus naming conventions
Currently the spapr PCI host bridge always uses the qdev core default for
naming the PCI bus hanging off it.  That is, the bus will be named
"<bridge id>.<bus num>".

This doesn't match the behaviour of a number of common (x86) PCI bridge
types, both host bridges and P2P bridges: e.g. 'ioh3420' and 'pci-bridge'.
For those, if the bridge has an id set, the bus name is set to the same
value without an added number.

This changes spapr_pci to follow the same convention.  Strictly, this is an
incompatible change in behaviour.  However, so far libvirt - the most
common client - doesn't use PAPR host bridges beyond the default one, and
anyone manually setting ids is at least as likely to be confused by our
difference from x86 convention as they are by the change from past
behaviour.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>